backend/requirements.txt
Flask==3.1.0
flask-cors
flask-sockets
gevent
gevent-websocket
pandas==2.2.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
dlib==19.24.6
Pillow==11.0.0
python-dotenv
google.generativeai
google-generativeai
flask-limiter==3.5.0
flask_sockets
colorama==0.4.6
requests

# Testing and db
pytest==7.4.3
pytest-flask==1.3.0
flask_sqlalchemy
mysql-connector==2.2.9
mysqlclient==2.0.3
pymysql==1.1.0
mysql-connector-python==8.0.33

# Face Recognition
numpy==1.23.5
face-recognition==1.3.0
mtcnn==0.1.1
tensorflow==2.16.2
pickleshare==0.7.5
pickle-mixin==1.0.2

psutil
uuid
platforms
fastapi
asyncio
jwt